refactor(journal/utils): extract helper functions
