+++ src/journal/utils/send-journal.ts
@@ -1,4 +1,4 @@
-import {Journal} from "./journal.type";
+import {Journal} from "../journal.type";
 import {Response} from "express";
 export const sendJournal = (res: Response) => {
     return (journal: Journal): void => {
